@@ -17,8 +17,8 @@ module.exports = function(config) {
       clearContext: false, // leave Jasmine Spec Runner output visible in browser
     },
     coverageIstanbulReporter: {
-      dir: require('path').join(__dirname, '../coverage'),
-      reports: ['html', 'lcovonly'],
+      dir: require('path').join(__dirname, './coverage'),
+      reports: ['html', 'lcovonly', 'text-summary'],
       fixWebpackSourcePaths: true,
     },
     reporters: ['mocha', 'kjhtml'],
@@ -34,5 +34,6 @@ module.exports = function(config) {
       },
     },
     singleRun: false,
+    restartOnFileChange: true,
   });
 };
